Fixed PL Expiration block

+++ xServer.D/Feature/PriceLock/PriceFeature.cs
@@ -244,7 +244,7 @@ namespace x42.Feature.PriceLock
 
                     if (price < 42000000)
                     {
-                        var expirationBlock = Convert.ToInt16(networkFeatures.BestBlockHeight + Convert.ToInt16(priceLockRequest.ExpireBlock));
+                        var expirationBlock = Convert.ToInt32(networkFeatures.BestBlockHeight + Convert.ToInt32(priceLockRequest.ExpireBlock));
 
                         using (X42DbContext dbContext = new X42DbContext(databaseSettings.ConnectionString))
                         {
